requirements.txt
PIL
Django==1.5.1
Pillow==2.0.0
argparse==1.2.1
distribute==0.6.24
django-braces==1.0.0
django-crispy-forms==1.3.2
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-jsonview==0.2.1
filebrowser-safe==0.2.27
grappelli-safe==0.2.22
html5lib==0.95
oauthlib==0.4.2
pytz==2013b
requests==1.2.3
requests-oauthlib==0.3.2
six==1.3.0
wsgiref==0.1.2
